requirements.txt
numpy<2.4
pyyaml
tqdm
torch==2.4.1
torchvision==0.19.1